Mastodon_Harvestor/mastodon_harvestor/requirements.txt
anyio==3.6.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
comm==0.1.2
emoji==2.2.0
ibm-cloud-sdk-core==3.16.5
ibmcloudant==0.4.1
idna==3.4
Mastodon.py==1.8.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pygments==2.14.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
